+++ Editor/PlayMakerUtilsDefine.cs
@@ -17,8 +17,19 @@ public class PlayMakerUtilsDefines
 {
 	static PlayMakerUtilsDefines()
 	{
-		PlayMakerDefines.AddScriptingDefineSymbolToAllTargets("PLAYMAKER_UTILS");
-		PlayMakerDefines.AddScriptingDefineSymbolToAllTargets("PLAYMAKER_UTILS_1_4_OR_NEWER");
+
+		#if ! PLAYMAKER_UTILS
+			PlayMakerEditorUtils.MountScriptingDefineSymbolToAllTargets("PLAYMAKER_UTILS");
+			
+		#endif
+
+		#if ! PLAYMAKER_UTILS_1_4_OR_NEWER
+			PlayMakerEditorUtils.UnMountScriptingDefineSymbolToAllTargets("PLAYMAKER_UTILS_1_4_OR_NEWER");
+		#endif
+
+		#if ! PLAYMAKER_UTILS_1_5_OR_NEWER
+			PlayMakerEditorUtils.UnMountScriptingDefineSymbolToAllTargets("PLAYMAKER_UTILS_1_5_OR_NEWER");
+		#endif
 	}
 
 }