+++ Changelog.md
@@ -7,6 +7,10 @@
 
 **new**: debug optional parameter for `PlayMakerEvent` `sendevent()` call
 
+**new**: Editor Utils to copy string to clipboard `PlayMakerEditorUtils.CopyTextToClipboard()`
+
+**Fix**: Fixed mounting and unmounting scripting symbols, properly based on PlaymakerDefines.cs. 
+
 ###1.4.2
 **Release Date:** 22/06/2017