+++ Changelog.md
@@ -11,6 +11,8 @@
 
 **Fix**: Fixed mounting and unmounting scripting symbols, properly based on PlaymakerDefines.cs. 
 
+**new**: `PLAYMAKER_UTILS_1_5_OR_NEWER` symbol
+
 ###1.4.2
 **Release Date:** 22/06/2017  
 
@@ -25,7 +27,7 @@
 ###1.4
 **Release Date:** 01/06/2017
 
-**new**: PlayMakerUtilsDefine class to feature PLAYMAKER\_UTILS and PLAYMAKER\_UTILS\_1\_4\_OR\_NEWER  
+**new**: PlayMakerUtilsDefine class to feature `PLAYMAKER_UTILS` and `PLAYMAKER_UTILS_1_4_OR_NEWER`  
 
 **new**: MainCamerTarget Public Class with property drawer like the owner class for proxy component creation